more ux chnages
changed signin to have its own screen, show up on first load as first screen to alllow users to sign in, create account, or go as guest. subsequent starts of app will send to loading screen and no longe ask to sign in.

Profile link goes to sign in if the user is "guest" otherwise goes to account settings.